+++ LuckyoneUI/Core/Core.lua
@@ -393,11 +393,13 @@ function Private:HandleToons()
 		['Player-4454-060E3657'] = true, -- [H] Druid
 		['Player-4454-060E364E'] = true, -- [H] Priest
 		['Player-4454-060E361A'] = true, -- [H] Shaman
+	} or Private.isTBC and {
+		-- (6412: Spineshatter)
+		['Player-6412-028A3A6D'] = true, -- [H] Hunter
+		['Player-6412-0336641F'] = true, -- [H] Priest
+		['Player-6412-02A39E0E'] = true, -- [H] Warlock
+		['Player-6412-02BBE8AB'] = true, -- [H] Hunter 2
 	} or Private.isClassic and {
-		-- (6112: Spineshatter)
-		['Player-6112-028A3A6D'] = true, -- [H] Hunter
-		['Player-6112-02A2F754'] = true, -- [H] Priest
-		['Player-6112-02A39E0E'] = true, -- [H] Warlock
 		-- (5827: Living Flame)
 		['Player-5827-0273D732'] = true, -- [A] Hunter
 		['Player-5827-0273D63E'] = true, -- [A] Paladin