+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-﻿## Version 3.30 [04-12-2025]
+﻿## Version 3.32 [14-01-2026]
+
+**All WoW Versions:**  
+- ElvUI 14.06+ compatibility  
+
+**Retail:**  
+- Removed WeakAuras import section in preparation for Midnight  
+
+**TBC:**  
+- Initial support for Burning Crusade Classic Anniversary  
+
+## Version 3.31 [24-12-2025]
+
+**All WoW Versions:**  
+- Bugfixes, no other changes  
+
+## Version 3.30 [04-12-2025]
 
 **All WoW Versions:**  
 - ElvUI 14.04+ compatibility